var refactor

+++ 20190606/FizzBuzz/src/main/java/com/company/FizzBuzz.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class FizzBuzz {
     public static void main(String[] args) {
-        for (int i = 1; i <= 100; i++) {
+        for (byte i = 1; i <= 100; i++) {
             if (i % 15 == 0) {
                 System.out.println("FIZZBUZZ");
             } else if (i % 5 == 0) {
